==> 04- Compile dataset from re-annotation pipeline.ipynb <==
path_data = '/'

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns 
import warnings
warnings.filterwarnings("ignore")
import os

import pandas as pd

import torch
from pkg_resources import packaging
from PIL import Image
import tqdm
from pathlib import Path
print("Torch version:", torch.__version__)

from PIL import ImageFile
ImageFile.LOAD_TRUNCATED_IMAGES = True

plt.style.use('ggplot')

data_unfiltered = pd.read_csv("/".join([path_data, "images_v2.csv"]))
#display(data_unfiltered)

len(data_unfiltered)

### Remove subjective topics

subjective_topics = ['Favorite home decorations', 'Favourite item in kitchen', 'Favourite sports clubs', 'How the most loved item is used', 'icons', 'Idols', 'Latest furniture bought', ' looking over the shoulder', 'Most loved item', 'Most loved toy', 'Most played songs on the radio', 'Music idol', 'Next big thing you are planning to buy', 'Playing with most loved toy', 'Thing I dream about having', 'Things I wish I had', 'Using most loved item', 'Youth culture', 'What I wish I could buy']

unique_topics = data_unfiltered['topics'].unique()

Subjective_topic_list = []

for topic in unique_topics:
    topic_ = topic.replace('/ ', ', ')
    topic_split = topic_.split(',')
    for split in topic_split:
        if split in subjective_topics:
            Subjective_topic_list.append(topic)

data = data_unfiltered[~data_unfiltered['topics'].isin(Subjective_topic_list)]
display(data)

len(data)

### Split topics

from collections import Counter
list_topics = list(data['topics'])


separate_topics = [t.lower().strip() for topic in list_topics for t in topic.split(",") ]


print(len(separate_topics))
#print(Counter(separate_topics))
set_topics = list(set(separate_topics))
print(len(set_topics))

# map each topic to list of corresponding images
dict_topic2img = {}
for list_topics, image_path in zip(data['topics'], data['imageRelPath']):
    for topic in list_topics.split(","):
        topic = topic.lower().strip()
        if topic not in dict_topic2img:
            dict_topic2img[topic] = set() #### here a set was used in place of list to avoid duplicate where keyword appears twice in a topic
        dict_topic2img[topic].add(image_path)

print(len(dict_topic2img))

#sanity checks 

ground_truth_counts = [len(dict_topic2img[i]) for i in dict_topic2img]
print(sum(ground_truth_counts))

topics = dict_topic2img.keys()

len(topics)

# Create prompts for the text encoder

temp = "This is a photo of "
English_topics = [temp + topic for topic in topics]

#English_topics

topic2english = {}
for topic, prompt in zip(topics, English_topics):
    topic2english[topic] = prompt

#topic2english

topic2english['family snapshots']

english_dict_topic2imgGT = {}

for topic in topics:
    key = topic2english[topic]
    english_dict_topic2imgGT[key] = dict_topic2img[topic]
print(len(english_dict_topic2imgGT))

english_dict_topic2imgGT.keys()

## Compiling dataset

import pickle
with open("data/dict_genfunction_topic2imgGT.pkl","rb") as f:
    generated_functions_dict_topic2imgGT = pickle.load(f)

len(generated_functions_dict_topic2imgGT)

topics_list = []
image_ids = []
functions_list = []

for topic_, images in dict_topic2img.items():
    topics_list += [topic_] * len(list(images))
    image_ids += images
    

#check if both dicts align
# list out keys and values separately
a_key_list = list(dict_topic2img.keys())
a_val_list = list(dict_topic2img.values())

# list out keys and values separately
b_key_list = list(generated_functions_dict_topic2imgGT.keys())
b_val_list = list(generated_functions_dict_topic2imgGT.values())

a_val_list == b_val_list

functions_list = []
for topic_, images in generated_functions_dict_topic2imgGT.items():
    functions_list += [topic_] * len(list(images))

    
len(functions_list)

import pickle
with open("image2Llavacaption_dict.pkl", 'rb') as f:
    image2Llavacaption_dict = pickle.load(f)

len(image2Llavacaption_dict)

# list(image2Llavacaption_dict.keys())[0].split()

llava_caption_list = []
list_of_paths = [path_data + s for s in image_ids]
for image_id in list_of_paths:
    llava_cap = image2Llavacaption_dict[image_id]
    llava_caption_list.append(llava_cap.split(':')[-1].lstrip())


import pickle
with open("image2BLIPcaption_dict.pkl", 'rb') as f:
    image2BLIPcaption_dict = pickle.load(f)

len(image2BLIPcaption_dict)

blip_caption_list = []
#list_of_paths = [path_data + s for s in image_ids]
for image_id in list_of_paths:
    blip_cap = image2BLIPcaption_dict[image_id]
    blip_caption_list.append(blip_cap)


print(f"Length of topics: {len(topics_list)}")
print(f"Length of image ids: {len(image_ids)}")
print(f"Length of generated functions: {len(functions_list)}")
print(f"Length of Llava captions: {len(llava_caption_list)}")
print(f"Length of BLIP captions: {len(blip_caption_list)}")

dataset_v1 = pd.DataFrame()
dataset_v1['topics'] = topics_list
dataset_v1['generated functions'] = functions_list
dataset_v1['Llava captions'] = llava_caption_list
dataset_v1['BLIP captions'] = blip_caption_list
dataset_v1['image ids'] = image_ids

dataset_v1.head()

dataset_v1.to_csv('data/dollar_street_reannotated_dataset_v1.csv')

with open ("llava_gpt4_generated_obj_dict.pkl", "rb") as f:
    llava_gpt4_generated_obj_dict = pickle.load(f)
    

len(llava_gpt4_generated_obj_dict)    

image_ids[0]

object_list = []
for funct, im_id in tqdm.tqdm(zip(functions_list, image_ids)):
    key = f"{funct}_{path_data}{im_id}"
    object_list.append(llava_gpt4_generated_obj_dict[key])
    

len(object_list)

dataset_v2 = pd.DataFrame()
dataset_v2['topics'] = topics_list
dataset_v2['generated functions'] = functions_list
dataset_v2['identified object'] = object_list
dataset_v2['Llava captions'] = llava_caption_list
dataset_v2['BLIP captions'] = blip_caption_list
dataset_v2['image ids'] = image_ids

dataset_v2.head()

dataset_v2.to_csv('data/dollar_street_reannotated_dataset_v2.csv')

## Add income

dataset_v2 = pd.read_csv('data/dollar_street_reannotated_dataset_v2.csv')

id_list = dataset_v2['image ids'].tolist()
income_list = [data[data['imageRelPath'] == i]['income'].item() for i in id_list]

len(income_list)
len(id_list)

data[data['imageRelPath'] == 'assets/5d4be06acf0b3a0f3f3392fa/5d4be06acf0b3a0f3f3392fa.JPG']['income'].item()

income_list[0]

dataset_v2['income'] = income_list

dataset_v2.to_csv('data/dollar_street_reannotated_with_income_v2.csv')

## Add scores

import pickle
with open('data/dict_default_topic2score.pkl', 'rb') as f:
    dict_default_topic2score = pickle.load(f)

len(dict_default_topic2score)

#dict_default_topic2score['This is a photo of armchair']

temp_dict = {}
for key, value in tqdm.tqdm(dict_default_topic2score.items()):
    for values in value:
        topic = key.replace("This is a photo of ","")
        score = values[0]
        img_id= values[1]
    
        new_key = f"{topic}_{img_id}"
        temp_dict[new_key] = score
    

len(temp_dict)

score_list = []
for topic, im_id in tqdm.tqdm(zip(dataset_v2['topics'].to_list(), dataset_v2['image ids'].to_list())):
    key = f"{topic}_{im_id}"
    score_list.append(temp_dict[key])
    

dataset_v2['scores'] = score_list

dataset_v2.head()

dataset_v2.to_csv('data/dollar_street_reannotated_with_income_v2.csv')
