==> 06- CLIP association test.ipynb <==
import os
import pandas as pd
from PIL import Image
from tqdm import tqdm
from pathlib import Path
import pickle
import torch
import clip
from transformers import CLIPProcessor, CLIPModel
from sentence_transformers import SentenceTransformer, util
import numpy as np
import gc

# Choose computation device
device = "cuda:0" if torch.cuda.is_available() else "cpu" 

# Load pre-trained CLIP model
model, preprocess = clip.load("ViT-B/32", device=device, jit=False)
model.cuda().eval()

path_data = '/'  # Base path for images

# Load the CSV file
data_unfiltered = pd.read_csv("/".join([path_data, "images_v2.csv"]))

len(data_unfiltered)

## Remove subjective topics

subjective_topics = ['Favorite home decorations', 'Favourite item in kitchen', 'Favourite sports clubs', 'How the most loved item is used', 'icons', 'Idols', 'Latest furniture bought', ' looking over the shoulder', 'Most loved item', 'Most loved toy', 'Most played songs on the radio', 'Music idol', 'Next big thing you are planning to buy', 'Playing with most loved toy', 'Thing I dream about having', 'Things I wish I had', 'Using most loved item', 'Youth culture', 'What I wish I could buy']

unique_topics = data_unfiltered['topics'].unique()

Subjective_topic_list = []

for topic in unique_topics:
    topic_ = topic.replace('/ ', ', ')
    topic_split = topic_.split(',')
    for split in topic_split:
        if split in subjective_topics:
            Subjective_topic_list.append(topic)

data = data_unfiltered[~data_unfiltered['topics'].isin(Subjective_topic_list)]
data = data[data.id != '5d4befbfcf0b3a0f3f353c2e'] #remove rows with corrupted image 
display(data)

len(data)

labels=["poor", "lower-middle", "upper-middle", "rich"]

#list_of_topic2prompt_dict[1]
data["quartile"] = pd.qcut(data["income"], q=[0, 0.25, 0.5, 0.75, 1],    
                         labels=labels)

## Split topics

from collections import Counter
list_topics = list(data['topics'])


separate_topics = [t.lower().strip() for topic in list_topics for t in topic.split(",") ]


print(len(separate_topics))
#print(Counter(separate_topics))
set_topics = list(set(separate_topics))
print(len(set_topics))

# map each topic to list of corresponding images
dict_topic2img = {}
for list_topics, image_path in zip(data['topics'], data['imageRelPath']):
    for topic in list_topics.split(","):
        topic = topic.lower().strip()
        if topic not in dict_topic2img:
            dict_topic2img[topic] = set() #### here a set was used in place of list to avoid duplicate where keyword appears twice in a topic
        dict_topic2img[topic].add(image_path)

print(len(dict_topic2img))

ground_truth_counts = [len(dict_topic2img[i]) for i in dict_topic2img]
print(sum(ground_truth_counts))

topics = dict_topic2img.keys()

with open("mapping_topic2function.pkl", 'rb') as f:
    topic2function = pickle.load(f)

dict_function2imgGT = {}

for topic in topics:
    key = topic2function[topic]
    dict_function2imgGT[key] = dict_topic2img[topic]
print(len(dict_function2imgGT))

dict_img2topic, dict_img2country, dict_img2incomelevel, dict_country2region, dict_img2income = {}, {}, {}, {}, {}
for list_topics, image_path, country, incomelevel, region, income in zip(data['topics'], data['imageRelPath'], data['country.name'], data['quartile'], data['region.id'], data['income']):
    dict_img2topic[image_path] = list_topics
    dict_img2country[image_path] = country
    dict_img2incomelevel[image_path] = incomelevel
    dict_country2region[country] = region
    dict_img2income[image_path] = income
    

#dict_topic2img

image_list = []
for imgs in dict_topic2img.values():
    image_list += imgs
len(image_list)

topic_list = []
for t, imgs in dict_topic2img.items():
    topics = [t] * len(imgs)
    topic_list += topics
len(topic_list)

function_list = [topic2function[f] for f in topic_list]
len(function_list)

country_list = [dict_img2country[c] for c in image_list]
income_list = [dict_img2incomelevel[i] for i in image_list]
region_list = [dict_country2region[cntry] for cntry in country_list]
income = [dict_img2income[inc] for inc in image_list]

functions = [' '.join(function.split(' ')[5 :]) for function in function_list]
function_topic = [f"the {clean_id} represents {function}" for clean_id, function in zip(topic_list, functions)]


data_sep = pd.DataFrame()
data_sep['topics'] = topic_list
data_sep['image ids'] = image_list
data_sep['generated functions'] = function_list
data_sep['income level'] = income_list
data_sep['income'] = income
data_sep['country'] = country_list
data_sep['continent'] = region_list
data_sep['function_topic'] = function_topic

data_sep.to_csv('one_image_many_topics.csv')

data = data_sep

path_images = [path_data + s for s in list(data['image ids'])]
image_ids1 = ['/'.join(s.split("/")[8:]) for s in path_images]

image_ids1[0].split("/")[1]

len(path_images)

image_path_list = []
imgs_corrupted = []
for image_path in tqdm(path_images):
        path_to_img = image_path
        if Path(path_to_img).is_file():
            image_path_list.append(path_to_img)
        else:
            imgs_corrupted.append(path_to_img)
len(image_path_list)

def load_image(url_or_path):
    if url_or_path.startswith("http://") or url_or_path.startswith("https://"):
        return Image.open(requests.get(url_or_path, stream=True).raw)
    else:
        return Image.open(url_or_path).convert("RGB")

count = 5000
end = len(image_path_list) // count + 1
print(end)

list_img_emb = []
image_names_list = []
for k in tqdm(range(0, end)):
    print(k)                    
    imgs_corrupted = []
    images_prep = []
    k_images = image_path_list[count*k:count*(k+1)]
    print(len(k_images))
    

    for image_path in k_images:
        if Path(image_path).is_file():
            image = Image.open(image_path).convert("RGB")
            images_prep.append(preprocess(image))
        
            
    images_prep = torch.stack(images_prep).to("cuda")
    torch.cuda.empty_cache()
    
    with torch.no_grad():
        image_features = model.encode_image(images_prep).float()
        image_features /= image_features.norm(dim=-1, keepdim=True)
        torch.cuda.empty_cache()
    list_img_emb.append(image_features)
    
    image_names = ['/'.join(image_path.split("/")[8:])  for image_path in image_path_list]
    image_names_list.append(image_names)
    
print("Image encoding completed..")

img_features = torch.cat(list_img_emb, dim=0)
img_features.shape

img_features.shape

# f = open("sep_clip_img_embedding.pkl","wb")

# # write the python object (dict) to pickle file
# pickle.dump(img_features,f)

# # close file
# f.close()

### Text and Image Sim

with open("sep_clip_img_embedding.pkl","rb") as f:
    img_embedding = pickle.load(f)
img_embedding.shape

img_features = img_embedding

gc.collect()
torch.cuda.empty_cache()

#### Change prompts here

text = data['function_topic'].to_list()


len(text)

count = 5000
end = len(text) // count + 1
print(end)
score_list = []

for k in tqdm(range(10)):
    k_img_features = img_features[count*k:count*(k+1)]
    k_text = text[count*k:count*(k+1)]
    
    with torch.no_grad():
        text_tokens = clip.tokenize(k_text).cuda()
        text_features = model.encode_text(text_tokens).float()
        text_features /= text_features.norm(dim=-1, keepdim=True)

    similarity = text_features.cpu().numpy() @ k_img_features.cpu().numpy().T # dot product of image and text features
    similarity_diag = similarity.diagonal()
    
    score_list += list(similarity_diag)


len(data)

len(score_list)

exp_1_results = data

exp_1_results["CLIP score"] = score_list

#exp_1_results = exp_1_results.drop(['Unnamed: 0'], axis=1)

exp_1_results.to_csv('exp_1_sep_function_topic_results.csv')

exp_1_results.head()
