==> 08- Performance gap Analysis.ipynb <==
import pandas as pd
import os
import pickle
from collections import Counter

import pickle
with open("data/topic_dict_topic2FNimg.pkl","rb") as f:
    topic_dict_topic2FNimg = pickle.load(f)

len(topic_dict_topic2FNimg)

performance_data = pd.read_csv('data/function_wise.csv')

topic_list = performance_data['topic']

 Aligning topic names

a = set(list(topic_dict_topic2FNimg.keys())) - set(topic_list)

set(topic_list)- set(list(topic_dict_topic2FNimg.keys()))

topic_dict_topic2FNimg['drainage'] = topic_dict_topic2FNimg[' drainage']
topic_dict_topic2FNimg['reading light'] = topic_dict_topic2FNimg[' reading light']
topic_dict_topic2FNimg['seeing the back of book'] = topic_dict_topic2FNimg[' seeing the back of book']
topic_dict_topic2FNimg['smells)'] = topic_dict_topic2FNimg[' smells)']
topic_dict_topic2FNimg['sofa'] = topic_dict_topic2FNimg[' sofa']

for t in a:
    topic_dict_topic2FNimg.pop(t)

len(topic_dict_topic2FNimg)

dict_topic2FNimg = topic_dict_topic2FNimg

Calaculating recall for topics

data = pd.read_csv('data/clean_data.csv')

# map each image to corresponding topics, countries, income level ... TODO
dict_img2topic, dict_img2country, dict_img2incomelevel, dict_country2region = {}, {}, {}, {}
for list_topics, image_path, country, incomelevel, region in zip(data['topics'], data['imageRelPath'], data['country.name'], data['quartile'], data['region.id']):
    dict_img2topic[image_path] = list_topics
    dict_img2country[image_path] = country
    dict_img2incomelevel[image_path] = incomelevel
    dict_country2region[country] = region

print(len(dict_img2topic))

# map each topic to list of corresponding images
dict_topic2img = {}
for list_topics, image_path in zip(data['topics'], data['imageRelPath']):
    for topic in list_topics.split(","):
        topic = topic.lower().strip()
        if topic not in dict_topic2img:
            dict_topic2img[topic] = set() #### here a set was used in place of list to avoid duplicate where keyword appears twice in a topic
        dict_topic2img[topic].add(image_path)

print(len(dict_topic2img))

list_all_incomelevelsGT = []
dict_topic2GTIncomelevels = {}
dict_all_incomelevels_GT = {}
for topic in dict_topic2img:
    list_imagesGT = dict_topic2img[topic]
    list_incomelevelsGT = [dict_img2incomelevel[img] for img in list_imagesGT]
    for incomelevel in list_incomelevelsGT:### Topics where Poor has higher Recall than other classes (Appendix)
        list_all_incomelevelsGT.append(incomelevel)
    #print(f"topic: {topic}; {Counter(list_countriesFN)}")
    dict_topic2GTIncomelevels[topic] = Counter(list_incomelevelsGT)
    
#print(f"ALL topics: {Counter(list_all_incomelevelsGT)}")
dict_all_incomelevels_GT = Counter(list_all_incomelevelsGT)

dict_topic2GTrichvspoor = {}
for topic, values in dict_topic2GTIncomelevels.items():
    temp_dict = {}
    temp_dict['low'] = values['poor'] + values['lower-middle']
    temp_dict['high'] = values['rich'] + values['upper-middle']
    
    dict_topic2GTrichvspoor[topic] = temp_dict
len(dict_topic2GTrichvspoor)

list_all_incomelevelsFN = []
dict_topic2FNIncomelevels = {}
dict_all_incomelevels_FN = {}
for topic in dict_topic2FNimg:
    list_imagesFN = dict_topic2FNimg[topic]
    list_incomelevelsFN = [dict_img2incomelevel[img] for img in list_imagesFN]
    for incomelevel in list_incomelevelsFN:
        list_all_incomelevelsFN.append(incomelevel)
    #print(f"topic: {topic}; {Counter(list_countriesFN)}")
    dict_topic2FNIncomelevels[topic] = Counter(list_incomelevelsFN)
    
#print(f"ALL topics: {Counter(list_all_incomelevelsFN)}")
dict_all_incomelevels_FN = Counter(list_all_incomelevelsFN)

dict_topic2FNrichvspoor = {}
for topic, values in dict_topic2FNIncomelevels.items():
    temp_dict = {}
    temp_dict['low'] = values['poor'] + values['lower-middle']
    temp_dict['high'] = values['rich'] + values['upper-middle']
    
    dict_topic2FNrichvspoor[topic] = temp_dict
len(dict_topic2FNrichvspoor)

Recall_perincomelevel_pertopic = {}
for topic in dict_topic2FNrichvspoor:
    FN_counter_ = dict_topic2FNrichvspoor[topic]
    GT_counter_ = dict_topic2GTrichvspoor[topic]
        
    temp_recall_ratio_ = Counter({key : 1 - FN_counter_[key] / GT_counter_[key] for key in FN_counter_ if sum(GT_counter_.values()) * 0.01 <= GT_counter_[key] })
    Recall_perincomelevel_pertopic[topic] = temp_recall_ratio_

t_list = topic_list
topic_high = [round(Recall_perincomelevel_pertopic[topic]['high'], 2) for topic in t_list]
topic_low = [round(Recall_perincomelevel_pertopic[topic]['low'], 2) for topic in t_list]



performance_data['topic_low'] = topic_low
performance_data['topic_high'] = topic_high


performance_data = performance_data.drop(columns = ['Unnamed: 0'])
performance_data.head()

performance_data['gap_topic'] = performance_data['topic_high'] - performance_data['topic_low']
performance_data['gap_function'] = performance_data['function_high'] - performance_data['function_low']
performance_data.head()

performance_data['delta_gap'] = performance_data['gap_topic'] - performance_data['gap_function']

performance_data.head()

sum(performance_data['delta_gap'])/len(performance_data['delta_gap'])

performance_data['delta_gap'].median()

from scipy.stats import wilcoxon

delta_gap = performance_data['delta_gap']

# Wilcoxon signed-rank
stat, p_value = wilcoxon(delta_gap, alternative='greater')
print(f"Median gap reduction = {delta_gap.median():.3f}, p = {p_value:}")

from scipy.stats import ranksums

ranksums(list(performance_data['gap_topic']), list(performance_data['gap_function']), alternative='greater')
